Pass second test version
